crop-prediction-backend/requirements.txt
numpy
pandas
joblib
Pillow
matplotlib
opencv-python
scikit-learn
seaborn
